requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
decorator==4.0.11
filetype==1.0.5
idna==3.2
imageio==2.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
moviepy==0.2.3.2
numpy==1.21.3
oauthlib==3.1.1
Pillow==8.4.0
pygame==2.0.1
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
tqdm==4.11.2
tweepy==3.10.0
urllib3==1.26.7
Werkzeug==2.0.1